@@ -34,7 +34,7 @@
     "clean:autogen": "find -type d -name autogen -exec rm -rfv '{}' \\; || true",
     "clean:full": "yarn clean:libndist && yarn clean:autogen",
     "format": "prettier --write \"{*,**/*}.{ts,tsx,js,jsx,json,yml,yaml,md}\"",
-    "lint": "eslint --ignore-path .gitignore \"./packages/**/*.{ts,tsx,js,jsx}\"",
+    "lint": "NODE_OPTIONS='--max-old-space-size=4096' eslint",
     "typecheck": "lerna run typecheck",
     "prepare-disabled": "lerna run prepare && husky install",
     "precommit": "yarn generate && lerna run --concurrency 1 --stream precommit && yarn lint-staged",
@@ -60,11 +60,14 @@
   "devDependencies": {
     "@apollo/rover": "^0.23.0",
     "@composedb/cli": "^0.7.1",
+    "@eslint/compat": "^1.1.1",
+    "@eslint/js": "^9.7.0",
     "@graphql-codegen/cli": "^5.0.2",
     "@graphql-codegen/client-preset": "^4.3.2",
     "@graphql-codegen/typescript-graphql-request": "^6.2.0",
     "@graphql-codegen/typescript-resolvers": "^4.2.1",
     "@graphql-codegen/typescript-urql": "^4.0.0",
+    "@types/eslint__js": "^8.42.3",
     "@types/jest": "^29.2.1",
     "@types/node": "^20.8.6",
     "@types/react": "^18.0.21",
@@ -74,26 +77,17 @@
     "@typescript-eslint/parser": "^7.16.1",
     "concurrently": "7.0.0",
     "env-cmd": "10.1.0",
-    "eslint": "^8.39.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-import-resolver-typescript": "^3.5.5",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jest": "^27.2.1",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint": "^9.7.0",
     "hasura-cli": "2.23.0",
     "husky": "7.0.4",
     "jest": "^29.2.2",
     "lerna": "^8.1.6",
-    "lint-staged": "12.3.5",
+    "lint-staged": "^15.2.7",
     "prettier": "2.7.1",
     "standard-version": "9.3.2",
     "ts-jest": "^29.0.3",
-    "typescript": "^5.0.0",
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^7.16.1",
     "wait-on": "6.0.1"
   },
   "resolutions": {
