@@ -1,4 +1,4 @@
-module.exports = {
+export default   {
   extends: ['plugin:@next/next/recommended'],
   rules: {
     'no-console': [
