@@ -1,9 +1,9 @@
-{
-  "rules": {
+export default {
+  rules: {
     "no-unused-vars": "off",
     "@typescript-eslint/no-unused-vars": "warn"
   },
-  "overrides": [
+  overrides: [
     {
       "files": ["./src/**/*.{ts,tsx,js,jsx}", "./stories/**/*.{ts,tsx,js,jsx}"],
       "rules": {
